Add socket rooms.
Rather than sending ALL tweets to a client, and having the client sort out
which tweets it should show, depending on its category, we now have the client
connect to a 'room'/category. We can then send new tweets to all clients connected to that room. Brilliant! Assuming it works as I think it does.