Workaround default-theme loading bug
If no theme is specified in the yaml, there's a problem somewhere loading the 'default' theme from the settings. This results in there being no (news/bbcradio) categories shown on the index page, and nothing can be selected, no tweets shown. Also nothing is shown if you go straight to the /news page. Here's a workaround.