Fix problem with long tweets from the Stream API.
Tweets that were longer than 140 characters because, for example,
they include the URL of an image at the end, which is no longer
counted towards the character limit, were showing in truncated
form when fetched via the Streaming API. The initial cached ones,
via the REST API, were fine. Now they all are.